Simplify project sttructure.
