+++ Availability/Sources/AvailabilityInfo.swift
@@ -32,3 +32,13 @@ public struct AvailabilityInfo {
   public let name: String
   public let modules: [ModuleInfo]
 }
+
+public protocol ModuleInfo {
+  var name: String { get }
+  var components: [ComponentInfo] { get }
+}
+
+public protocol ComponentInfo {
+  var name: String { get }
+  var availability: [String: Bool] { get }
+}
